+++ packages/core/src/Table/stories/TableRenderers/ProgressColumnRenderer.tsx
@@ -1,0 +1,115 @@
+import React, { useMemo, useState } from "react";
+import {
+  HvCellProps,
+  HvPagination,
+  HvRowInstance,
+  HvTable,
+  HvTableBody,
+  HvTableCell,
+  HvTableContainer,
+  HvTableHead,
+  HvTableHeader,
+  HvTableRow,
+  hvProgressColumn,
+  useHvData,
+  useHvPagination,
+} from "@hitachivantara/uikit-react-core";
+
+import { EmptyRow, NewRendererEntry, makeRenderersData } from "../storiesUtils";
+
+export const ProgressColumnRenderer = () => {
+  const columns = useMemo(() => {
+    return [
+      hvProgressColumn<NewRendererEntry, string>(
+        {
+          Header: "Probability",
+          accessor: "riskScore",
+          style: { width: 125 },
+          disableSortBy: true,
+          id: "probability-header",
+        },
+        (row) => row.original.riskScore,
+        () => 100,
+        "secondary"
+      ),
+    ];
+  }, []);
+
+  const initialData = useMemo(() => makeRenderersData(64), []);
+
+  const [data] = useState(initialData);
+
+  const {
+    getTableProps,
+    getTableBodyProps,
+    prepareRow,
+    headers,
+    page,
+    getHvPaginationProps,
+  } = useHvData<NewRendererEntry, string>(
+    {
+      columns,
+      data,
+      defaultColumn: {
+        Cell: ({ value }: HvCellProps<NewRendererEntry, string>) =>
+          value ?? "—",
+      },
+    },
+    useHvPagination
+  );
+
+  const rowRenderer = (pages: HvRowInstance<NewRendererEntry, string>[]) => {
+    return pages.map((row, index) => {
+      prepareRow(row);
+
+      return (
+        <React.Fragment key={row.id}>
+          <HvTableRow
+            {...row.getRowProps({
+              "aria-rowindex": index + 1,
+            })}
+          >
+            {row.cells.map((cell) => (
+              <HvTableCell {...cell.getCellProps()}>
+                {cell.render("Cell")}
+              </HvTableCell>
+            ))}
+          </HvTableRow>
+        </React.Fragment>
+      );
+    });
+  };
+
+  return (
+    <>
+      <HvTableContainer>
+        <HvTable
+          {...getTableProps({
+            "aria-rowcount": data.length,
+          })}
+          style={{
+            width: 230,
+          }}
+        >
+          <HvTableHead>
+            <HvTableRow>
+              {headers.map((col) => (
+                <HvTableHeader
+                  {...col.getHeaderProps()}
+                  key={col.Header}
+                  id={col.id}
+                >
+                  {col.render("Header")}
+                </HvTableHeader>
+              ))}
+            </HvTableRow>
+          </HvTableHead>
+          <HvTableBody {...getTableBodyProps()}>
+            {page.length === 0 ? <EmptyRow height={100} /> : rowRenderer(page)}
+          </HvTableBody>
+        </HvTable>
+      </HvTableContainer>
+      <HvPagination {...getHvPaginationProps?.()} />
+    </>
+  );
+};